requirements.txt
aiogram>=3.4.1
SQLAlchemy>=2.0.28
aiosqlite>=0.20.0
python-dotenv>=1.0.1
